Split data like images and videos into new repository.
Downloading or git cloning RespiRate from the web took too long
due to the sample.mp4 video and images in data/ that, for most
users, would have been unnecessary to run RespiRate. We've split
this data into its own web repository so it will not
automatically be downloaded.
This changes all the in-code references to this data to our new
data repo, https://github.com/MDA-Courtyard/RespiRate_data.